Revert "Reland "goma: use vpython3 for upload_goma_logs.py""
This reverts commit 8030e5c6455b7ca140d45189268c6d0bcb74366a.

Reason for revert:
https://ci.chromium.org/ui/p/chromium/builders/ci/fuchsia-x64-cast/164588/overview

Original change's description:
> Reland "goma: use vpython3 for upload_goma_logs.py"
>
> This reverts commit 400dd70537a5e40863354faef3ac45465b5479b9.
>
> Reason for revert:
> .vpython3 is added to .gitattribute for recipe bundler.
>
> Original change's description:
> > Revert "goma: use vpython3 for upload_goma_logs.py"
> >
> > This reverts commit e3e4aa8173cffd21c0c72591cdb7f31b47248885.
> >
> > Reason for revert:
> > https://ci.chromium.org/ui/p/chromium/builders/ci/fuchsia-x64-cast/164281/overview
> >
> > Original change's description:
> > > goma: use vpython3 for upload_goma_logs.py
> > >
> > > This is to fix
> > > https://ci.chromium.org/ui/p/chromium/builders/ci/Mac%20Builder%20(py2%20less)/1294/overview
> > >
> > > Recipe-Nontrivial-Roll: build_limited
> > > Recipe-Nontrivial-Roll: chrome_release
> > > Bug: 1327534
> > > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3690033
> > > Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
> >
> > Bug: 1327534
> > No-Presubmit: true
> > No-Tree-Checks: true
> > No-Try: true
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3692484
> > Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
> > Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
> > Auto-Submit: Takuto Ikuta <tikuta@chromium.org>
> > Owners-Override: Takuto Ikuta <tikuta@chromium.org>
>
> Bug: 1327534
> Fixed: 1334049
> Recipe-Nontrivial-Roll: build_limited
> Recipe-Nontrivial-Roll: chrome_release
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3692485
> Commit-Queue: Takuto Ikuta <tikuta@chromium.org>

Bug: 1327534
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3697557
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: Takuto Ikuta <tikuta@chromium.org>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>

+++ recipes/upload_goma_logs.py
@@ -98,7 +98,7 @@ def main():
 
   # For CompileEvents.
   parser.add_argument(
-      '--build-id', default=0, type=int, help='unique ID of the current build'
+      '--build-id', default=0, type=long, help='unique ID of the current build'
   )
   parser.add_argument(
       '--build-step-name', default='', help='step name of the current build'
